+++ src/main/java/ru/magzook/sellersrestservice/exception/TransactionWithIdNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.magzook.sellersrestservice.exception;
 
 public class TransactionWithIdNotFoundException extends EntityWithIdNotFoundException {
-    public TransactionWithIdNotFoundException(int id) {
+    public TransactionWithIdNotFoundException(long id) {
         super("transaction", id);
     }
 }
